+++ sumo-logic/release_notes.md
@@ -1,2 +1,14 @@
 ## What's Fixed
-- Fixed the Search Job related functions/actions to use cookies as per Sumo Logic API reference document.
+- Insights search parameter (q)
+
+## What's New
+- Update the Severity of an Insight
+- Update the Status of an Insight
+- Retrieves the history for a specific insight by its ID
+- Retrieves the comments for a specific insight by its ID.
+- Retrieves the enrichments for a specific insight by its ID.
+- Get an Insight's List of Involved Entities
+- Ingestion playbooks
+
+## Note:
+- "MAC Address" and "User Agent" indicator types must be added to the default 'Indicators Types' picklist